requirements.txt
amqp==1.4.9
anyjson==0.3.3
APScheduler==3.7.0
asgiref==3.3.4
billiard==3.3.0.23
cached-property==1.5.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
distro==1.5.0
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-heroku==0.3.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.5.0
joblib==1.0.1
kombu==3.0.37
numpy==1.19.5
pandas==0.25.3
prompt-toolkit==3.0.3
psycopg2==2.9.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sqlparse==0.4.1
tabula-py==2.2.0
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.1
